shared: Extract modules with handle-related types
